Modifies index page to use section tags
This commit modifies the index page to use section tags instead of div
tags. This cleans up the HTML. Additionally, added some functionality to
display project lists when each project category is hovered over. This
functionality still needs to be worked out as its not behaving as it
should just yet.