Adds function to remove duplicates from scroller
The way in which the sroller works is by duplicating the items of the
skills list once the about page is loaded. This was creating a problem
because each time the about section was navigated away from and then
back to, the items would be duplicated again, causing the scroller to
double in speed every time. This commit adds a function to remove
duplicates from the list before the items are re-duplicated.