Fixes mobile layout of about menu section
This commit fixes the layout of the mobile version of the about menu.
The scroller works and is moving horiziontally as desired. The entire
menu is now visible and the user can scroll the bio section to see the
overflow of the text, with the skills scroller being fixed to the bottom
of the screen.

Things that need to be fixed:

- The mask is not working on the scroller
- There seemse to be a descrepency in the spacing of the items within
the scroller.
- Fix the project menu similarly, so that the project-category and
project-lists are visible in their entirety, and there is a scroller for
the project descriptions.