Fixes formatting of the project menu on mobile
The project menu will now be displayed within on page on mobile, with
the project description menu being scrollable.

Look into the possibility of making the project list smaller, or being
able to resize it. For most phones the current size is fine, but for
devices less than about 700px high, the project description is too
small.

Any easy fix would be to resize for mobile devices with a height of less
than 700px.