Adds functionality for content type in ci_scripts
The upload_files_to_dev.py script willl now detect the content type of
the file being uploaded and add it to the object when being uploaded to
the bucket. If the content type is not detected, a message will be
displayed stating that the content type must be added manually for that
object.