Adds workflow to push to prod
Adds a workflow that pushes the changes to the producation bucket. It
also invalidates the CloudFront cache so that the changes can be viewed
immediately. The upload script was modified slightly so that environment
variables can be passed in from the workflow files and it can be used
for both dev and prod buckets. Its name was changed to reflect this
change.