Loads images when project menu is clicked
This commit implements a better version of lazy loading. Instead of
loading each project image when the project description is clicked, all
of the images are loaded when the project menu is clicked. This is
achieved by using a javascript function to change the loading attritube
to eager when the project menu is clicked. This way, images are loaded
and the animation always looks smooth